backend/requirements.txt
fastapi[standard]
uvicorn
sqlmodel
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
python-dotenv
cors
